Source Code/requirements.txt
timm
einops
tqdm
numpy
matplotlib
Pillow
opencv-python
thop
scikit-learn
scipy
scikit-image
huggingface_hub